feat(frontend): multiple price charts with browser persistence
Turn the single-product price view into a dashboard of charts, one per
product. Extract the chart into a ProductChart component; the page keeps a
list of selected products and renders a chart card for each, with a remove
button. The list is saved to localStorage and restored on load, so charts
persist across browser sessions.
