+++ frontend/app/pages/prices.vue
@@ -1,20 +1,4 @@
 <script setup lang="ts">
-import {
-  Chart as ChartJS,
-  CategoryScale,
-  LinearScale,
-  PointElement,
-  LineElement,
-  Title,
-  Tooltip,
-  Legend,
-  TimeScale
-} from 'chart.js'
-import 'chartjs-adapter-date-fns'
-import { Line } from 'vue-chartjs'
-
-ChartJS.register(CategoryScale, LinearScale, PointElement, LineElement, Title, Tooltip, Legend, TimeScale)
-
 useSeoMeta({ title: 'Price History · cartwatch' })
 
 const supabase = useSupabaseClient()
@@ -22,38 +6,27 @@ const supabase = useSupabaseClient()
 type ProductHit = { id: string; name: string; short_name: string; unit: string | null }
 
 const query = ref('')
-const selectedProduct = ref<ProductHit | null>(null)
 const searchResults = ref<ProductHit[]>([])
 const searching = ref(false)
-const loadingHistory = ref(false)
 
-type PricePoint = {
-  date: string
-  unit_price: number
-  store_chain: string
-  raw_name: string
-  quantity: number
-  item_unit: string
-  currency: string
-}
-const history = ref<PricePoint[]>([])
+// Persisted list of charts (one per product), kept in this browser.
+const STORAGE_KEY = 'cartwatch:price-charts'
+const products = ref<ProductHit[]>([])
 
-const CHAIN_COLORS: Record<string, string> = {
-  REWE: '#e2001a',
-  Lidl: '#0050aa',
-  'Aldi Süd': '#00519e',
-  'Aldi Nord': '#003087',
-  Edeka: '#f5a800',
-  Penny: '#c8102e',
-  Netto: '#ffd100',
-  Kaufland: '#e2001a',
-  dm: '#d40511',
-  Rossmann: '#e2001a',
-  Amazon: '#ff9900',
-  Other: '#94a3b8'
-}
+onMounted(() => {
+  try {
+    const saved = localStorage.getItem(STORAGE_KEY)
+    if (saved) products.value = JSON.parse(saved)
+  } catch {
+    // ignore malformed storage
+  }
+})
 
-const CURRENCY_SYMBOLS: Record<string, string> = { EUR: '€', USD: '$', GBP: '£' }
+watch(products, (val) => {
+  if (import.meta.client) localStorage.setItem(STORAGE_KEY, JSON.stringify(val))
+}, { deep: true })
+
+const addedIds = computed(() => new Set(products.value.map(p => p.id)))
 
 async function searchProducts() {
   if (query.value.length < 2) {
@@ -75,126 +48,15 @@ async function searchProducts() {
   searching.value = false
 }
 
-async function selectProduct(product: ProductHit) {
-  selectedProduct.value = product
-  query.value = product.name
+function addProduct(product: ProductHit) {
+  if (!addedIds.value.has(product.id)) products.value.push(product)
+  query.value = ''
   searchResults.value = []
-  await loadHistory(product.id)
 }
 
-async function loadHistory(productId: string) {
-  loadingHistory.value = true
-  history.value = []
-
-  const { data } = await supabase
-    .from('receipt_items')
-    .select(`
-      raw_name,
-      quantity,
-      unit_price,
-      units ( symbol ),
-      receipts!inner (
-        purchased_at,
-        currency,
-        stores!inner (
-          store_chains!inner ( name )
-        )
-      )
-    `)
-    .eq('canonical_product_id', productId)
-    .order('receipts(purchased_at)', { ascending: true })
-
-  if (data) {
-    history.value = data.map((row: any) => ({
-      date: row.receipts.purchased_at,
-      unit_price: row.unit_price,
-      store_chain: row.receipts.stores.store_chains.name,
-      raw_name: row.raw_name,
-      quantity: row.quantity,
-      item_unit: row.units?.symbol ?? '',
-      currency: row.receipts.currency
-    }))
-  }
-  loadingHistory.value = false
+function removeProduct(id: string) {
+  products.value = products.value.filter(p => p.id !== id)
 }
-
-const currencySymbol = computed(() => {
-  const c = history.value[0]?.currency ?? 'EUR'
-  return CURRENCY_SYMBOLS[c] ?? `${c} `
-})
-
-// Distinct units actually present across the receipt items for this product.
-const distinctUnits = computed(() =>
-  [...new Set(history.value.map(p => p.item_unit).filter(Boolean))]
-)
-
-// Prefer the canonical product's unit; fall back to the item unit only when unambiguous.
-const displayUnit = computed(() =>
-  selectedProduct.value?.unit ?? (distinctUnits.value.length === 1 ? distinctUnits.value[0]! : null)
-)
-
-// True when there's no canonical unit and items disagree — prices aren't comparable yet.
-const unitsMixed = computed(() =>
-  !selectedProduct.value?.unit && distinctUnits.value.length > 1
-)
-
-const yAxisLabel = computed(() =>
-  `Price (${currencySymbol.value}${displayUnit.value ? ` / ${displayUnit.value}` : ''})`
-)
-
-const chartData = computed(() => {
-  const chains = [...new Set(history.value.map(p => p.store_chain))]
-  return {
-    datasets: chains.map(chain => ({
-      label: chain,
-      data: history.value
-        .filter(p => p.store_chain === chain)
-        .map(p => ({ x: new Date(p.date).getTime(), y: p.unit_price, meta: p })),
-      borderColor: CHAIN_COLORS[chain] ?? '#94a3b8',
-      backgroundColor: (CHAIN_COLORS[chain] ?? '#94a3b8') + '33',
-      tension: 0.3,
-      pointRadius: 4,
-      pointHoverRadius: 6
-    }))
-  }
-})
-
-const chartOptions = computed(() => ({
-  responsive: true,
-  maintainAspectRatio: false,
-  interaction: { mode: 'nearest' as const, intersect: true },
-  scales: {
-    x: {
-      type: 'time' as const,
-      time: { unit: 'month' as const },
-      title: { display: true, text: 'Date' }
-    },
-    y: {
-      title: { display: true, text: yAxisLabel.value },
-      ticks: { callback: (v: any) => `${currencySymbol.value}${Number(v).toFixed(2)}` }
-    }
-  },
-  plugins: {
-    legend: { position: 'bottom' as const },
-    tooltip: {
-      callbacks: {
-        // Full original receipt name as the tooltip heading.
-        title: (items: any[]) => items[0]?.raw?.meta?.raw_name ?? '',
-        label: (ctx: any) => {
-          const m = ctx.raw.meta as PricePoint
-          const unit = displayUnit.value ?? m.item_unit
-          return `${ctx.dataset.label}: ${currencySymbol.value}${ctx.parsed.y.toFixed(2)}${unit ? ` / ${unit}` : ''}`
-        },
-        afterLabel: (ctx: any) => {
-          const m = ctx.raw.meta as PricePoint
-          const date = new Date(m.date).toLocaleDateString()
-          const qty = `Qty: ${m.quantity}${m.item_unit ? ` ${m.item_unit}` : ''}`
-          return [qty, date]
-        }
-      }
-    }
-  }
-}))
 </script>
 
 <template>
@@ -204,14 +66,14 @@ const chartOptions = computed(() => ({
         Price History
       </h1>
       <p class="text-muted text-sm">
-        Search for a product to see how its price has evolved across stores.
+        Add a chart per product to compare price trends across stores. Your charts are saved in this browser.
       </p>
     </div>
 
     <div class="relative mb-8">
       <UInput
         v-model="query"
-        placeholder="Search products… e.g. milk, eggs"
+        placeholder="Search products to add a chart… e.g. milk, eggs"
         icon="i-lucide-search"
         size="lg"
         class="w-full"
@@ -225,58 +87,35 @@ const chartOptions = computed(() => ({
         <button
           v-for="product in searchResults"
           :key="product.id"
-          class="w-full text-left px-4 py-2.5 hover:bg-elevated flex items-center gap-3 transition-colors"
-          @click="selectProduct(product)"
+          class="w-full text-left px-4 py-2.5 hover:bg-elevated flex items-center gap-3 transition-colors disabled:opacity-50"
+          :disabled="addedIds.has(product.id)"
+          @click="addProduct(product)"
         >
           <UIcon name="i-lucide-package" class="text-muted shrink-0" />
           <span>{{ product.name }}</span>
-          <span class="text-muted text-sm ml-auto">{{ product.short_name }}</span>
+          <span
+            v-if="addedIds.has(product.id)"
+            class="ml-auto text-xs text-primary flex items-center gap-1"
+          >
+            <UIcon name="i-lucide-check" /> added
+          </span>
+          <span v-else class="text-muted text-sm ml-auto">{{ product.short_name }}</span>
         </button>
       </div>
     </div>
 
-    <div v-if="loadingHistory" class="flex justify-center py-16">
-      <UIcon name="i-lucide-loader-circle" class="animate-spin text-3xl text-muted" />
+    <div v-if="products.length" class="space-y-6">
+      <ProductChart
+        v-for="product in products"
+        :key="product.id"
+        :product="product"
+        @remove="removeProduct(product.id)"
+      />
     </div>
 
-    <template v-else-if="selectedProduct && history.length">
-      <div class="mb-3 flex flex-wrap items-center gap-2">
-        <UBadge variant="subtle">
-          {{ history.length }} data points
-        </UBadge>
-        <span class="text-muted text-sm">for <strong>{{ selectedProduct.name }}</strong></span>
-        <UBadge
-          v-if="displayUnit && !unitsMixed"
-          color="neutral"
-          variant="outline"
-        >
-          {{ currencySymbol }} / {{ displayUnit }}
-        </UBadge>
-        <UBadge
-          v-if="unitsMixed"
-          color="warning"
-          variant="subtle"
-          icon="i-lucide-triangle-alert"
-        >
-          Mixed units: {{ distinctUnits.join(', ') }} — prices not comparable
-        </UBadge>
-      </div>
-      <div class="h-80 border border-default rounded-xl p-4 bg-background">
-        <Line :data="chartData" :options="chartOptions" />
-      </div>
-    </template>
-
-    <div
-      v-else-if="selectedProduct && !loadingHistory"
-      class="text-center py-16 text-muted"
-    >
-      <UIcon name="i-lucide-inbox" class="text-4xl mb-2" />
-      <p>No price history found for <strong>{{ selectedProduct.name }}</strong>.</p>
-    </div>
-
-    <div v-else-if="!selectedProduct" class="text-center py-16 text-muted">
-      <UIcon name="i-lucide-trending-up" class="text-4xl mb-2" />
-      <p>Search for a product above to see its price history.</p>
+    <div v-else class="text-center py-16 text-muted">
+      <UIcon name="i-lucide-line-chart" class="text-4xl mb-2" />
+      <p>Search for a product above to add your first chart.</p>
     </div>
   </UContainer>
 </template>
